update: sidebar 다른 페이지에도 적용
