+++ frontend/client/src/components/SideBar/RecentGoodsView.spec.tsx
