+++ frontend/client/src/components/SideBar/RecentGoodsView.tsx
