+++ frontend/client/src/pages/CategoryGoods/CategoryGoods.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import styled from 'styled-components';
 import { useParams } from '@src/lib/CustomRouter';
 import CategoryGoodsList from '@src/pages/CategoryGoods/CategoryGoodsList/CategoryGoodsList';
+import SideBar from '@src/components/SideBar/SideBar';
+import Footer from '@src/components/Footer/Footer';
 
 const appendQuotationMarks = (text: string) => `"${text}"`;
